Update to have code in actual solution
